replace a single element in a immutable list of type class
