+++ src/main/kotlin/Player.kt
@@ -1,0 +1,7 @@
+package org.example
+
+data class Player(val id: Int,
+                  val matchesPlayed: Int,
+                  val totalRuns: Int,
+                  val name: String,
+                  val team: String)
