+++ src/main/kotlin/ReplaceAll.kt
@@ -1,10 +1,6 @@
 package org.example
 
-data class Player(val id: Int,
-    val matchesPlayed: Int,
-    val totalRuns: Int,
-    val name: String,
-    val team: String)
+
 
 fun main(){
   val playerList = listOf<Player>(
@@ -17,5 +13,5 @@ fun main(){
     updatedPlayers.forEach { player-> println(player)}
 }
 fun replaceTeam(list: List<Player>, newTeam: String) : List<Player> {
-     return list.map { player -> player.copy(team = newTeam)  }
+     return list.map { player -> player.copy(team = newTeam)}
 }
